FEATURE: Cache info about program

+++ app/src/main/java/ua/tools/escondido/tvprogram/services/impl/InternalStorageCache.java
@@ -1,4 +1,4 @@
-package ua.tools.escondido.tvprogram.utils;
+package ua.tools.escondido.tvprogram.services.impl;
 
 
 import android.content.Context;
@@ -10,31 +10,35 @@ import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.util.Date;
 
-public class InternalStorage {
-    private InternalStorage() {}
+import ua.tools.escondido.tvprogram.services.Cacheable;
+import ua.tools.escondido.tvprogram.utils.DateUtils;
 
-    public static void writeObject(Context context, String key, Object object) throws IOException {
+public class InternalStorageCache<E> implements Cacheable<E>{
+
+    @Override
+    public void writeObject(Context context, String key, E data) throws IOException {
         FileOutputStream fos = context.openFileOutput(key, Context.MODE_PRIVATE);
         ObjectOutputStream oos = new ObjectOutputStream(fos);
-        oos.writeObject(object);
+        oos.writeObject(data);
         oos.close();
         fos.close();
     }
 
-    public static Object readObject(Context context, String key) throws IOException,
-            ClassNotFoundException {
-        cleanUpData(context);
+    @Override
+    public E readObject(Context context, String key, Boolean isCleanUpNeeded) throws IOException, ClassNotFoundException {
+        if(isCleanUpNeeded) {
+            cleanUpData(context);
+        }
         FileInputStream fis = context.openFileInput(key);
         ObjectInputStream ois = new ObjectInputStream(fis);
-        Object object = ois.readObject();
-        return object;
+        return (E) ois.readObject();
     }
 
-    public static void cleanUpData(Context context){
+    protected static void cleanUpData(Context context){
         String today = DateUtils.formatChannelAccessDate(new Date(System.currentTimeMillis() - 18000000));
         String nextDay = DateUtils.formatChannelAccessDate(new Date(System.currentTimeMillis() + 86400000));
         for(String fileName : context.fileList()){
-            if(!(fileName.contains(today) || fileName.contains(nextDay))){
+            if(!(fileName.contains(today) || fileName.contains(nextDay)) && (!fileName.startsWith("_"))){
                 context.deleteFile(fileName);
             }
         }
